backend/requirements.txt
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.11
python-decouple==3.8
Pillow==12.1.1
requests==2.32.3
